Create ContactTeamCard folder and its respective files
